week-09/day-02/guardians-of-the-galaxy/src/main/java/com/greenfox/guardiansofthegalaxy/controller/DraxController.java
package com.greenfox.guardiansofthegalaxy.controller;

import com.greenfox.guardiansofthegalaxy.model.FoodDTO;
import com.greenfox.guardiansofthegalaxy.service.DraxService;
import java.util.List;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/drax")
public class DraxController {

  @Autowired
  private DraxService draxService;

  @GetMapping("/foods")
  public List<FoodDTO> findAll() {
    return draxService.findAll();
  }

}
